requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
click==8.1.7
gunicorn==21.2.0
python-dotenv==1.0.0
tabulate==0.9.0
